Generic styles for input/button/select
Made a generic constant with general styling for all input types that can then be overwritten in specific scenarios instead.

Made the media dropdown use a generic dropdown styling instead (not much changed either way)

Changed the method for displaying borders when focusing on an element. Now it hides it by default by making it transparrent.