Fix searching (#518)
* Added date-fns

Great library for handling date formating

* Delete package-lock.json

We use yarn, not npm. Somehow it snuck in my previous PR

* Update yarn.lock

* Added/modified styled components

Gave search results a flexbox and made it generally look more pleasing

* Function to figure out season

Put this into a logical component. Thinking this file could be used for any type of date functions.

* Update SearchMedia.tsx

Added new elements for displaying results

* Fixed searching

An earlier commit broke searching where it wouldn't use the content in the searchbox in the query. This change fixes that.

* Remove search when no content

Removes search results if the user removes the search query