Move fragment outside of query
Moved GraphQL queries into the types folder so that we keep all queries in one place.

Moved fragment into its own .fragment.graphql file.

Modified my prettier config to be consistent with previous practices which means some of the files I previously modified with this will be changed back. Will do the rest of those in the next commit