Refactor (#532)
* Configure prettier formating for consistency

* Function declaration for functional components related to search

I'm only doing the ones related to search for now while the edit stuff is being worked on in a separate branch.

* Got rid of the necessity of graphql.macro

For now, only on the search/home page. It means one place to import all queries from and using the hooks generated by the codegen

* Move fragment outside of query

Moved GraphQL queries into the types folder so that we keep all queries in one place.

Moved fragment into its own .fragment.graphql file.

Modified my prettier config to be consistent with previous practices which means some of the files I previously modified with this will be changed back. Will do the rest of those in the next commit

* Prettier consistency

I reverted a prettier change in the previous commit. This is just reverting those changes back to how it was before

* updated .prettierignore

* Made changes to the rest of the GraphQL stuff

Moved graphql query/fragment files to /graphql/ instead of /src/types

Renamed query files to .query.graphql for consistency

Removed uneeded files

* More prettier changes